1. Shift Schedule deamon thread task timer into a same package. 2. Add in Byzantine-boardcast function to enhancement FileMetaInfo-synchronize under diskFile's crystalBlock cluster. 3. support crystalBlock cluster running under both internet & Lan  network mode.

+++ BoyJDiskFile/src/main/java/cn/Boy/DiskFile/dao/AppPlatformDocTypeRelationDao.java
@@ -41,9 +41,9 @@ public interface AppPlatformDocTypeRelationDao {
     public int deleteByGuid(String appdocTypeRelationGuid);
 
     @Update("update AppPlatformDocTypeRelation set isActive=0 where appdocTypeRelationID=#{appdocTypeRelationID}")
-    public int setToNotActive(int appdocTypeRelationID);
+    public int setToNotActive(long appdocTypeRelationID);
 
     @Update("update AppPlatformDocTypeRelation set isActive=1 where appdocTypeRelationID=#{appdocTypeRelationID}")
-    public int setToActive(int appdocTypeRelationID);
+    public int setToActive(long appdocTypeRelationID);
 
 }
